+++ src/schnetpack/data/__init__.py
@@ -1,0 +1,6 @@
+from .atoms import *
+from .definitions import *
+from .loader import *
+from .merging import *
+from .partitioning import *
+
