+++ src/schnetpack/data/definitions.py
@@ -1,0 +1,18 @@
+
+
+class Structure:
+    """
+    Keys to access structure properties in `schnetpack.data.AtomsData`
+    """
+    Z = '_atomic_numbers'
+    charge = '_charge'
+    atom_mask = '_atom_mask'
+    R = '_positions'
+    cell = '_cell'
+    pbc = '_pbc'
+    neighbors = '_neighbors'
+    neighbor_mask = '_neighbor_mask'
+    cell_offset = '_cell_offset'
+    neighbor_pairs_j = '_neighbor_pairs_j'
+    neighbor_pairs_k = '_neighbor_pairs_k'
+    neighbor_pairs_mask = '_neighbor_pairs_mask'
