+++ src/scripts/schnetpack_ani1.py
@@ -12,6 +12,7 @@ from torch.optim import Adam
 from torch.utils.data.sampler import RandomSampler
 
 import schnetpack as spk
+import schnetpack.data
 from schnetpack.datasets.ani1 import ANI1
 from schnetpack.utils import compute_params, to_json, read_from_json
 
@@ -304,9 +305,9 @@ if __name__ == '__main__':
     data_train, data_val, data_test = ani1.create_splits(*args.split, split_file=split_path)
 
     logging.info('load data...')
-    train_loader = spk.data.AtomsLoader(data_train, batch_size=args.batch_size, sampler=RandomSampler(data_train),
-                                        num_workers=4, pin_memory=True)
-    val_loader = spk.data.AtomsLoader(data_val, batch_size=args.batch_size, num_workers=2, pin_memory=True)
+    train_loader = schnetpack.data.AtomsLoader(data_train, batch_size=args.batch_size, sampler=RandomSampler(data_train),
+                                               num_workers=4, pin_memory=True)
+    val_loader = schnetpack.data.AtomsLoader(data_val, batch_size=args.batch_size, num_workers=2, pin_memory=True)
 
     if args.mode == 'train':
         logging.info('calculate statistics...')
@@ -338,8 +339,8 @@ if __name__ == '__main__':
         logging.info("...training done!")
     elif args.mode == 'eval':
         logging.info("evaluating...")
-        test_loader = spk.data.AtomsLoader(data_test, batch_size=args.batch_size,
-                                           num_workers=4, pin_memory=True)
+        test_loader = schnetpack.data.AtomsLoader(data_test, batch_size=args.batch_size,
+                                                  num_workers=4, pin_memory=True)
         with torch.no_grad():
             evaluate(args, model, train_args.property, train_loader, val_loader, test_loader, device)
         logging.info("... done!")
