+++ src/scripts/schnetpack_md17.py
@@ -13,6 +13,7 @@ from torch.optim import Adam
 from torch.utils.data.sampler import RandomSampler
 
 import schnetpack as spk
+import schnetpack.data
 from schnetpack.datasets import MD17
 from schnetpack.utils import compute_params, to_json, read_from_json
 
@@ -337,9 +338,9 @@ if __name__ == '__main__':
     data_train, data_val, data_test = md17.create_splits(*train_args.split, split_file=split_path)
 
     logging.info('load data...')
-    train_loader = spk.data.AtomsLoader(data_train, batch_size=args.batch_size, sampler=RandomSampler(data_train),
-                                        num_workers=4, pin_memory=True)
-    val_loader = spk.data.AtomsLoader(data_val, batch_size=args.batch_size, num_workers=2, pin_memory=True)
+    train_loader = schnetpack.data.AtomsLoader(data_train, batch_size=args.batch_size, sampler=RandomSampler(data_train),
+                                               num_workers=4, pin_memory=True)
+    val_loader = schnetpack.data.AtomsLoader(data_val, batch_size=args.batch_size, num_workers=2, pin_memory=True)
 
     if args.mode == 'train':
         logging.info('calculate statistics...')
@@ -370,8 +371,8 @@ if __name__ == '__main__':
         logging.info("...training done!")
     elif args.mode == 'eval':
         logging.info("evaluating...")
-        test_loader = spk.data.AtomsLoader(data_test, batch_size=args.batch_size,
-                                           num_workers=2, pin_memory=True)
+        test_loader = schnetpack.data.AtomsLoader(data_test, batch_size=args.batch_size,
+                                                  num_workers=2, pin_memory=True)
         evaluate(args, model, train_loader, val_loader, test_loader, device)
         logging.info("... done!")
     else:
